modificaciones

+++ ProyectoConMS/pruducto-ms/src/main/java/cl/fullstack/pruducto_ms/repository/ProveedorRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 import cl.fullstack.pruducto_ms.entity.ProveedorEntity;
 
 @Repository
-public interface ProveedorRepository extends JpaRepository<ProveedorEntity, Long> {
+public interface ProveedorRepository extends JpaRepository<ProveedorEntity, Integer> {
 }
